requirements.txt
streamlit
pandas
google-generativeai
streamlit-lightweight-charts
python-dotenv
streamlit-extras
ast
